listings/views_api.py
from rest_framework import viewsets
from .models import Property, Appointment
from .serializers import PropertySerializer, AppointmentSerializer

class PropertyViewSet(viewsets.ModelViewSet):
    queryset = Property.objects.all()
    serializer_class = PropertySerializer

class AppointmentViewSet(viewsets.ModelViewSet):
    queryset = Appointment.objects.all()
    serializer_class = AppointmentSerializer

