requirements.txt
Django==5.2.8
djangorestframework==3.16.0
dj-database-url==3.0.1
psycopg[binary]
gunicorn
Pillow
whitenoise
